Commit: Iniital Running Version
Explain:
This is the initial running version of the split app. I have created the first part of the app which is the users portion and setup the common file system that is going ot hold eveything tha tis used but the rest of the applications. I have also added a testing method just to make usre that it is runnning and working. I also setup the docker portion and the docker server is currently up and running at the time of this commit. Everything will be tested on docker including the changes that occur.