Commit: Added first method - Login
Explain:
I created the first view and html template test which is going to be the main screen users see when you go on split.com. this is a login view that will allow new users to register for this applicaiotn. I just want to make usre everything is working as i planned. I will also be testing the docker images and make they are still working with new changes.