Commit: Added Profile Functionality
Explain:
I have now added the ability for a created user to setup the first part of the profile which asks for general information that the application can use within itself as well as used to setup Synapse Account for verification before transfering money. Next is o add the profile verificaiton requirements that will be saved locally.