Commit: Added All Forms - User Account
Explain:
I have added all of the different forms that user has to setup before he can access his account and being using the application. As of right now NO record has been created or can be create within the views. THey simply display the forms. THe next step is to go ahead and begin integrating objects in the database and start manageing real objets and accounts in the database. Also adding a logout method to the views.py file so the user can logout.