Commit: Fixed Error - Updated Models
Explain:
There was an issue with importing a specific modeule within the applicaiton that was causing the docker image not to run properly. To fix this I installed django-localflavor on local machine as well as added it to requirements.txt. This install allows me to include things like state selection in the profile form as well as make sure that the zip code and phone number are in the correct format when they are submitted to the database. I did migrate the models file so I now have a database that was created, which will allow me to start adding actual objects to the database for testing the user setup and regisutraiton as well as login