Commit: User Login Signup and Logout function is working
Explain:
I have successfully integrated the signup system in split. You will enter specifcied credentials and then be directed to your home page. From there you can logout of your account and be automatically redirected to the login screen. In the login screen you can enter your credentials you used to create your account and get back logged into your account. Before I go in and clean up my code and comment everything, I wanted to test and see if a docker image of this current app passes and copies the existing database and all of the objects within it or a new database is build and new objects need to be created.