Commit:Django Login System Interation
Explain:
I havd integrated the login system that django has within its framework that makes it very easy to utilize within the application. When a user signs up or logs the login method is called and the user is passed after it is authenticated. Then for all of the methods and defs that needs a user to be logged into will have a simple @login_required tag at the top which will ensure that someone is logged in before accessing the application. The logout system is also very simple and easy way to logout... Next is to comment all of the code and clean up the code as well as add functionality.