Commit: Fixed Login Issue + Signup Issue
Explain:
I was getting an issue with the signup page, where if a user tries to sign up with a new account and there is a username that already exists in the database, It gives an invalid http respoesn. I corrected that by adding a little step that checks for error and displays errors if there are any. THis solved tat issue and now users will get alerts if there are any issues with the signup information. I also fixxed the view that allows users to properly login and logout of their account.. Next, I will comment all of the code and also do some clean up.