Commit: Added Users Profile
Explain:
I have now added a username search. This will allow the logged in user to search for other users in the database and if the username is valid, it will redirect the user to the searched users profile to see his info and eventually send a friend request to the searched user.