Commit: Sending User Requests Functionality
Explain:
I have added the ability for a user to send a friend request to another user by simply searching for the user an then sending a request from the searched users profile page. The person who receives the friend request will see it. THe one who send the friend request will not see the friend request....