Commit: Friends system integration
Explain:
I have no integrated a friendship system. Now a user can send another user a friend request. When a user receives a friend request from the sending user, he or she can accpet and decline the friend request. if the friend request is accepted, it will create a new record of the friendship and delete the initial friend request. If the request is declined, the inital request that was sent will be deleted and then the user will just be redirected to his or her profile.. The users home page will dipslay all the request they received as well as the list of friends who have accepted the friend request.