Commit: Added Group Creating Functionality
Explain:
I have added the creating group functionality. When a users goes to the group home page, he will see a button to create a new group. Once he clicks on the create group link, he will fill out a form displayed and select the group members from the list of his friends in the form. Once the form is submiited, it will create a group objects and then go through all of his friends and the ones he selected will be added to the group in the form of a member object that is going to be stored inthe database.