Commit: Bundles Created
Explain:
I have added the bundels function to the groups. THis allows a user to create a group of expenses that need to be paid into one simple transaction and transfer. You enter each item and its amount. THe amounts will be split abnd added up into a singel amount that each person will owe. It will be verified and transfered like a normal expense. I also ever so slightly restructured the settings page and users views.py file