Commti: Added all initial activity functionality
Explain:
I have implimented all of the required activities integration into the application that is needed. For the users activties, this includes creating the account, sending and accepting friend requests, and creating a group. The activties for the group include adding members, creating expenses, creating bundles. there will be more additions to the activities but this is the fundamental beggiing of the acitivities development