Commit: hosting and leaving
Explain:
I have now integrated the ability to switch a host from one user to another. A host is the one who can create a bundle or expense. I have also added the ability fro users to leave a group if they do not have any outstanding balances due and if they are not the current host. Otherwise they can leave the group and their record will be deleted as a member.