Commit: Added Synapse Integration
Explain:
I have now added and integrated an accounts tab. THis accounts tab will allow users to link their bank accounts to the application. With this link, you can transfer money from a bank acount to the uses split account to be transfered through different users. This is the core of the applicaiton. Right now you can link your bank account through account and routing number or ligin to your bank account and access all of your accounts. There is a lot of things that need to be added when going to a production version but for now this is sufficient.