Commit:Completed Refactorying and Synapse User Verification
Explain:
Completed all of split refactoring that was needed to be done. I have commented everything that is needed to be commented. I have integrated the synapse api to the point where user verification is the last part of the split and synapse integration. users can now properly link bank accounts to split and save info locally. Users are now verificed when they setup their profiles. Transacitons integration is the last that is needed to be done.