Commit: Transactions processig is implimented into the project
Explain:
I have now added the ability for a user to confirm a expense in a group and then initiate a transfer of funds between two users through their linked bank accounts. With the live version, one person needs to be verified in order ro process a transfer with real funds between both users synapse accounts.