Commit: Moved Friend Requst to Activity
Explain:
I changed friend requests from a hard friedn request on the users home to an activity objects that the user will use to accept or decline any and all freidn requests. THis is done so that I can have a single place where all of the users information is displayed as a block rather than two different blocks which make th application to split up rather than easy for hte user. I also started the add members to existing group but had an issue I am working to solve right now...